main features and errors handling

+++ controllers/cards.js
@@ -1,97 +1,81 @@
 const Card = require('../models/card');
 const {
-  CREATED, NOT_FOUND, BAD_REQUEST, SERVER_ERROR,
+  CREATED,
 } = require('../app');
+const NotFoundError = require('../errors/not-found-error');
+const ValidationError = require('../errors/validation-error');
+const ForbiddenError = require('../errors/forbidden-error');
 
-class CardNotFoundError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'CardNotFound';
-    this.statusCode = NOT_FOUND;
-  }
-}
-
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.send(cards))
-    .catch(() => res.status(SERVER_ERROR).send({ message: 'Произошла ошибка' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(CREATED).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' });
-        return;
+        return next(new ValidationError("Переданы некорректные данные при создании карточки"));
       }
-      res.status(SERVER_ERROR).send({ message: 'Произошла ошибка' });
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => {
-      throw new CardNotFoundError('Карточка с указанным _id не найдена');
-    })
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err instanceof CardNotFoundError) {
-        res.status(err.statusCode).send({ message: err.message });
-        return;
-      }
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId).orFail(() => {
+    throw new NotFoundError('Карточка с указанным _id не найдена');
+  }).then(card => {
+
+    if (JSON.stringify(card.owner._id) !== JSON.stringify(req.user._id)) {
+      throw new ForbiddenError("Недостаточно прав для выполнения операции");
+    }
+    Card.findByIdAndRemove(req.params.cardId)
+      .then((card) => res.send(card))
+      .catch((err) => { return err });
+  })
+    .catch(err => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные карточки' });
-        return;
+        return next(new ValidationError("Переданы некорректные данные карточки"));
       }
-      res.status(SERVER_ERROR).send({ message: 'Произошла ошибка' });
-    });
+      next(err);
+    })
 };
 
-module.exports.setCardLike = (req, res) => {
+module.exports.setCardLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
     .orFail(() => {
-      throw new CardNotFoundError('Передан несуществующий _id карточки');
+      throw new NotFoundError('Передан несуществующий _id карточки');
     })
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные при постановке лайка' });
-        return;
+        return next(new ValidationError("Переданы некорректные данные при постановке лайка"));
       }
-      if (err instanceof CardNotFoundError) {
-        res.status(err.statusCode).send({ message: err.message });
-        return;
-      }
-      res.status(SERVER_ERROR).send({ message: 'Произошла ошибка' });
+      next(err);
     });
 };
 
-module.exports.deleteCardLike = (req, res) => {
+module.exports.deleteCardLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
     .orFail(() => {
-      throw new CardNotFoundError('Передан несуществующий _id карточки');
+      throw new NotFoundError('Передан несуществующий _id карточки');
     })
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные при удалении лайка' });
-        return;
+        return next(new ValidationError("Переданы некорректные данные при постановке лайка"));
       }
-      if (err instanceof CardNotFoundError) {
-        res.status(err.statusCode).send({ message: err.message });
-        return;
-      }
-      res.status(SERVER_ERROR).send({ message: 'Произошла ошибка' });
+      next(err);
     });
 };
